push: handle push new job + disable_jobs_from_gui
Fixed the case where we try to push a new job to the server and the
jobs repository is configured with disable_jobs_from_gui = true.

In this case the server will return a 404 when we try to retrieve the
job's XML from it to inject the disabled flag in the to-be-pushed XML,
and we just have to ignore the error and push the local XML as-is.